Modified text of paper.
Introduced macro for U238.

In introduction, explained that we'll only present results for the 10.2cm collimator in the summary - other collimators in the final paper.

Added and tidied up todos.

Removed table about neutron yield - information to go in main text.

Completed text for the first part of the results section: neutron contour plot, neutron profile, neutron spectrum (with comparison to earlier results)